test-project/requirements.txt
norma-orm>=0.1.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
